+++ src/pyramid/request.py
@@ -1,5 +1,4 @@
 from collections import deque
-import json
 
 from zope.interface import implementer
 from zope.interface.interface import InterfaceClass
@@ -13,14 +12,17 @@ from pyramid.interfaces import (
     ISessionFactory,
 )
 
-from pyramid.compat import text_, bytes_, native_, iteritems_
-
 from pyramid.decorator import reify
 from pyramid.i18n import LocalizerRequestMixin
 from pyramid.response import Response, _get_response_factory
 from pyramid.security import AuthenticationAPIMixin, AuthorizationAPIMixin
 from pyramid.url import URLMethodsMixin
-from pyramid.util import InstancePropertyHelper, InstancePropertyMixin
+from pyramid.util import (
+    InstancePropertyHelper,
+    InstancePropertyMixin,
+    text_,
+    bytes_,
+)
 from pyramid.view import ViewMethodsMixin
 
 
@@ -226,10 +228,6 @@ class Request(
             return False
         return adapted is ob
 
-    @property
-    def json_body(self):
-        return json.loads(text_(self.body, self.charset))
-
 
 def route_request_iface(name, bases=()):
     # zope.interface treats the __name__ as the __doc__ and changes __name__
@@ -281,7 +279,7 @@ def call_app_with_subpath_as_path_info(request, app):
 
     # compute new_path_info
     new_path_info = '/' + '/'.join(
-        [native_(x.encode('utf-8'), 'latin-1') for x in subpath]
+        [text_(x.encode('utf-8'), 'latin-1') for x in subpath]
     )
 
     if new_path_info != '/':  # don't want a sole double-slash
@@ -328,7 +326,7 @@ def apply_request_extensions(request, extensions=None):
     if extensions is None:
         extensions = request.registry.queryUtility(IRequestExtensions)
     if extensions is not None:
-        for name, fn in iteritems_(extensions.methods):
+        for name, fn in extensions.methods.items():
             method = fn.__get__(request, request.__class__)
             setattr(request, name, method)