+++ CHANGES.rst
@@ -31,8 +31,9 @@ Features
   to ``add_route`` is the empty string (``''``).
   See https://github.com/Pylons/pyramid/pull/3420
 
-Bug Fixes
----------
+- No longer define ``pyramid.request.Request.json_body`` which is already
+  provided by WebOb. This allows the attribute to now be settable.
+  See https://github.com/Pylons/pyramid/pull/3447
 
 Deprecations
 ------------